Disable test for formula with tags; fix template for not using formulas with tags inside
Upgrading to ClosedXML 0.100 broke this functionality: previously it was possible to put a formula inside a cell and then to retrieve an original value it had (e.g. &="Total:"<<sum>>). Now this does not work (which is correct) and making it right would require somewhat major refactoring for which no one has time