Skip test SubrangesTests.Simple("Subranges_WithSort_tMD3.xlsx")
Caused by ClosedXML bug. When a sheet is copied (in this case to temp
sheet), it doesn't add links to calculation chain. But then, when it
tries to insert rows, formulas are shifted.

When XLRangeBase.InsertRowsAboveInternal is inserting rows, it also
shifts formulas of cells. It calls FormulaSlice.Set with shiufted
formula. The FormulaSlice then removes calc chain link for original
formula and adds for new formula. The problem is that it throws when
old formula doesn't have a link (and it doesn't because of copy).

ClosedXML should add links to calculation chain when sheet is copied,
but I am not sure. Should we even throws on missing link anyway?
ClosedXML will get a lot of files without correct calc chain.