Skip tag evaluation for non-text values
The bug in SubrangesTests.MultipleSubRanges was caused by setting a value
of a cell to string.Empty to option tag row that was later sorted as the
first row instead of last row.

This reduces that chance by parsing tags only from string values. Other
types can never contains <<tag>> text anyway, so it's fine.