Fix GroupTagTests.Simple(tLists3_options.xlsx | issue#111_autofilter_with_delete.xlsx)
Both tests had same two problems:
* A lot fo cells was maked as different, because in the gauge, they had
  default charset while the output had russina charset.
* The cell G24 was marked as unlocked, but in gauge it was locked

1. Style font char set difference
Charset was added to the Font in 0.104, so the equality comparer didn't
check it before, but does now. Therefore it is safe to ignore. Charset is
auxiliry property for a selection of a fallback font anyway.

2. G24 protection lock vs unlock

The problem was that cell G24 was locked in 0.102.2 and wasn't locked in
0.104.2. The root cause was also charset difference in a style of the
cell.

The ProtectedTag class first unlocks all used cells through

`ws.Cells().ForEach(c => { c.Style.Protection.Locked = false; });`

The ws.Cells() in returns only used cells. That is cells that don't have
value and have same style as a worksheet (relevant properties for this
case). That was true for the cell G24 in 0.102.2, because it ignored
charset property of a style. The cell G24 was thus evaluated as unused
and lock wasn't changed (although it was in fact locked, but so was sheet
- no difference in sheet vs cell style).

The 0.104.2 does however take into account a charset of a style. It found
that cell contains Default charset while sheet style contains Russian
charset. Therefore style of a cell was different and thus cell wasn't
considered empty.

ProtectedTag unlocks the non-empty cells (including G24) and that was the
root reason of discrepancy.

The gauge workbook can therefore be safely replaced by generated one.