Fix data due to stable sort->stable sort
Pre-0.103, the range sorting wasn't stable (it was a custom quick-sort). Post 0.103, the range sorting is stable (as it should be, just like Excel plus it makes sense). The sorting is used when values are separated into groups. Originally, the order of items wasn't kept, but now the order of items is kept.

The item groups in the test files are generallt orders and they have PK order id. As such, they are loaded from db in sequential order. Therefore the item groups should keep orderNo (PK) in ascending order,because fo stabel sort. The orderNo is *not* actually sorted, but because it was loaded from db in ascending order (thanks to PK), it looks sorted.

Also remove charset from font definition in styles.xml.

* GroupTagTests_DisableOutline.xlsx - sort orders in each group, remove charset from font
* GroupTagTests_FormulasInGroupRow.xlsx - sort orders in each group, remove charset from font, fix inconsistent border
* GroupTagTests_MergeLabels.xlsx - sort orders in each group, remove charset from font
* GroupTagTests_MergeLabels2.xlsx - sort orders in each group, remove charset from font
* GroupTagTests_PlaceToColumn.xlsx - sort orders in each group, remove charset from font
* GroupTagTests_SummaryAbove.xlsx - sort orders in each group, remove charset from font
* GroupTagTests_SummaryAbove.xlsx - sort orders in each group, remove charset from font, border didn't match in merged regions => check that matches (ignore border) and then replace original file.
* GroupTagTests_NestedGroups.xlsx - sort orders in each group, remove charset from font
* GroupTagTests_MultiRanges.xlsx - remove charset from font
* GroupTagTests_WithHeader.xlsx - sort orders in each group, remove charset from font
* GroupTagTests_Collapse.xlsx - sort orders in each group, remove charset from font