Fix SubrangesTests.SingleEmptySubsetCorrectSum
The test failed, because the original solution didn't clear out the <<sum>>
in the option row. The result thus had <<sum>> in the output instead of
empty cell.

I think the whole extra branch should be deleted, but that would require
deep modification of RangeTemplate and I don't know enought to do that.

Adjust solution for #251