Update pivot table generation example.
The pivot structure in 0.103 was completely overhauled and it is now
necessary to set the tabular layout.

Honestly, the whole "create pivot table" feature should be scrapped.

It's a wrong way to go about a generating a pivot table. The pivot table
should be created in Excel and should reference the range. That way, user
would have full control over all ascpets of the pivot table. The only thing
necessary would be to adjust the range and refresh the cache.

Instead, there is a special tag that contains a very limited subset of pivot
table options and the report generator tries to create new pivot table...
That's pointless.